# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AdCraft AI - Video Generation API
  description: |
    AI-powered video generation platform that transforms product images into professional commercials.
    
    ## Features
    - **AI Video Generation**: Create professional commercials from product images using Google Veo AI
    - **Interactive Chat**: Refine video concepts through natural language conversation
    - **Real-time Status**: Track video generation progress with WebSocket-like polling
    - **Multi-language Support**: Full English and Japanese localization
    - **Cost Management**: Built-in budget tracking and cost optimization
    - **Production Monitoring**: Comprehensive health checks and performance metrics
    
    ## Authentication
    Most endpoints are publicly accessible for the demo application.
    Admin endpoints require API key authentication via Bearer token.
    
    ## Rate Limits
    - Video Generation: 1 request per hour per IP, 3 per day maximum
    - Chat Refinement: 10 messages per session maximum
    - Status Checks: 1 request per second per job
    - Admin Endpoints: Requires valid admin API key
    
    ## Cost Information
    - Average cost per video generation: $1.81-2.01
    - Budget alerts at 50%, 75%, 90% thresholds
    - Real-time cost tracking available through admin dashboard
  version: 1.0.0
  contact:
    name: AdCraft AI Support
    url: https://github.com/your-org/adcraft-ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /status/{jobId}:
    get:
      tags:
        - Status Tracking
      summary: Get Video Generation Status
      description: |
        Retrieves the current status of a video generation job.
        Provides real-time updates on processing progress, cost tracking, and completion estimates.
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          description: Unique identifier for the video generation job
          schema:
            type: string
            example: "job_67890123"
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
              example:
                jobId: "job_67890123"
                status: "completed"
                progress: 100
                message: "Video generation completed successfully"
                videoUrl: "https://storage.googleapis.com/adcraft-videos/job_67890123.mp4"
                actualCost: 1.82
                processingTime: 387
                completedAt: "2025-08-28T10:34:27.000Z"
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error retrieving job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin API key for accessing administrative endpoints
  schemas:
...
    JobStatusResponse:
      type: object
      properties:
        jobId:
          type: string
          example: "job_67890123"
        status:
          type: string
          enum: [queued, processing, completed, failed]
          example: "completed"
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Progress percentage
          example: 100
        message:
          type: string
          description: Current status message
          example: "Video generation completed successfully"
        videoUrl:
          type: string
          format: uri
          description: URL to the generated video (when completed)
          example: "https://storage.googleapis.com/adcraft-videos/job_67890123.mp4"
        actualCost:
          type: number
          format: float
          description: Actual cost incurred
          example: 1.82
        processingTime:
          type: integer
          description: Processing time in seconds
          example: 387
        startedAt:
          type: string
          format: date-time
          example: "2025-08-28T10:28:00.000Z"
        completedAt:
          type: string
          format: date-time
          example: "2025-08-28T10:34:27.000Z"
        errorDetails:
          type: string
          description: Error details (when status is failed)
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Product name is required and must be between 1-100 characters"
        correlationId:
          type: string
          example: "corr_abc123"
        details:
          type: object
          description: Additional error details (development only)
...